backend/auth-service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pyjwt
bcrypt
pydantic
email-validator
